block_06_JWT_OAuth_OIDC/app/requirements.txt
amqp==5.3.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
billiard==4.2.2
celery==5.5.3
cffi==2.0.0
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
ecdsa==0.19.1
fastapi==0.121.0
greenlet==3.2.4
h11==0.16.0
idna==3.11
kombu==5.5.4
packaging==25.0
passlib==1.7.4
prompt_toolkit==3.0.52
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
python-dateutil==2.9.0.post0
python-jose==3.5.0
redis==7.0.1
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
asyncpg
pydantic[email]
python-multipart
